Feature/better traceability (#14)
* initial changes

* Dropped balance in account, fixed notifications

* Unified monetary units

* singleton checks

+++ aratiri-app/src/main/java/com/aratiri/transactions/application/processor/TransactionProcessor.java
@@ -3,10 +3,8 @@ package com.aratiri.transactions.application.processor;
 import com.aratiri.infrastructure.persistence.jpa.entity.TransactionEntity;
 import com.aratiri.transactions.application.dto.TransactionType;
 
-import java.math.BigDecimal;
-
 public interface TransactionProcessor {
-    BigDecimal process(TransactionEntity transaction);
+    long process(TransactionEntity transaction);
 
     TransactionType supportedType();
 }
